Add files via upload
Refactoring to update t-digest logging for latency and key metrics. Focus of app is on tracking temperature, and logging the power settings etc for t-digests as well.
The code will no longer move voltage and frequency for a target hashrate as this will flucutate due to bitcoin network activitiy. If you want to hash-harder, raise your temperature.